backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pydantic
python-multipart
